{{cookiecutter.project_slug}}/src/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_slug}}"
version = "0.1.0"
description = ""
authors = ["Dharwin Pérez <dhararon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.1"
Pillow = "^5.4"
django-environ = "^0.4.5"
djangorestframework = "^3.9"
factory-boy = "^2.11"
django-filter = "^2.1"
gunicorn = "^19.9"
django-storages = "^1.7"
django-cors-headers = "^2.4"
django-softdelete-it = "^0.1.0"
psycopg2 = "^2.8"

[tool.poetry.dev-dependencies]
pytest = "^4.1"
pytest-sugar = "^0.9.2"
django-debug-toolbar = "^1.11"
pytest-django = "^3.4"
pytest-picked = "^0.4.0"
pytest-tldr = "^0.1.5"
django-test-plus = "^1.1"
django_extensions = "^2.1"
pytest-factoryboy = "^2.0"
pre-commit = "^1.17"
django-test-without-migrations = "^0.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
